Alterados valores do endpoint media para GUID, para comodar utilização de GUID
